// types/v2.4.2/api.ts
export interface IncomingTransactionRequest {
  header: Omit<IncomingHeader, 'id' | 'received_at' | 'created_at' | 'updated_at'>;
  details: Omit<IncomingDetail, 'id' | 'header_id' | 'created_at' | 'updated_at'>[];
}